triplestore: Cleanup graphdb image
This commit cleans up the GraphDB Image. In particular it makes the
following improvements:

Uses a standalone healthcheck executable (written in golang) instead of
a curl-based approach. This removes the need to `apt-get install` in the
runtime container and is less likely to break in future releases.

Remove entrypoint.sh script in favor of using 'USER' command and
appropriatly chmoded volumes. Previously, the 'chmod' was performed in
the entrypoint script. This has been cleanly moved to image build time.